Move Utility.next_perm into FoldUtil
Our unit tests run in a node instance, so they can't pull in
any dependencies that are incompatible with node. The Utility class
contains a bunch of PIXI.Point-related functions.

The NuPACK folder uses this function, and we want to unittest nupack,
so the function gets moved to a node-compatible location.