TextInputObject
There's no great way to do this in PIXI/WebGL without rolling our
own, which would inevitably be a stupid amount of effort and would
be inferior to a native widget.

Instead, we create an <input> object on the DOM and float it over
the PIXI canvas. It's not on the display list, and will always be
rendered in front of anything in PIXI land.