Eliminate redundant use of Promise.resolve()
I found there were actually very few cases where this could be done.   Many of the uses of Promise.resolve() were not in an asynch function.

In EternaApp.setup(), it was used in an async lambda, but only one case of the switch statment used it, and the compiler didn't like the mixed metafor.    Since the rest of the lambda didn't use any await's, I just striped off the async and left the Promise.resolve().

In all, only one case where the Promise.resolve() could be eliminated was found.

Fixes: #219