Don't attempt to animate bases on mutate when in simple graphics mode (#680)
## Summary
When in simple graphics mode, if you mutate a base, it would slightly offset the base positions

## Implementation Notes
This was because we perform the same animation we use randomly over time on loops whenever you mutate a base. However, in low performance mode, a future update sets the animation to disabled, causing just a slight base movement. This patch checks if simple graphics mode is enabled before attempting to start animating

## Testing
Manual