Support variable barcode lengths (#685)
## Summary
Going forward, we are expecting to have synthesis runs that would have
more sequences than we could currently make unique barcodes for with
only a barcode of length 7. This takes the recent backend update adding
a barcode length field and uses it when determining the barcode indices,
falling back to the previous value of 7 if not specified.

## Implementation Notes
This logic mirrors the updated backend logic. I've opted to have the
barcode "count back" from the tail if a barcode start is not specified
in order to be more robust. Behavior if the barcode is placed past the
end of the RNA remains unhandled, and is considered a reasonable setup
error.

## Testing
Ensured proper validation behavior in the following scenarios:
* No barcode start or length, unused barcode
* No barcode start or length, used barcode
* Barcode length without start, unused barcode
* Barcode length without start, used barcode
* Barcode length and start, unused barcode
* Barcode length and start, used barcode

## Related Issues
This is a companion to commit 9cacd73a7cd6384341be4c2d978bb313ce8f1a8d
on the legacy website