Update WASM components to compile + load under newer emscripten versions (#686)
## Summary
This includes a handful of changes to account for recent(ish) changes in
Emscripten that caused issues when compiling with newer versions

## Implementation Notes
* In 3.1.27 (released 11/29/22), the default stack size was reduced from
5MB to 64KB. We need larger available stack space, so we revert this
setting
* In v1.38.41 (released 08/07/2019), callMain was no longer exported by
default. Only nupack exports a main function we intend to call, so we
enabled it specifically for that module
* `-Wimplicit-int` and `-Wint-conversion` were enabled, but cause errors
in vendor Vienna code. We revert those errors to warnings.
* In 1.39.16 (released 05/15/2020), Emscripten modules now provide an
actual promise instead of a "promise-like" interface, which we now use

## Testing
Basic sanity checks loading puzzles, unit tests