Support structure copy and paste target structure (#690)
## Summary
Using magic glue to specify a target structure can be painful if you
already have a particular target structure in mind that requires many
changes, especially if loading from an external tool. This allows for a
dot bracket target structure to be pasted directly.

Multiple users have also requested being able to copy structures, and
this will also help if wanting to make the target structure match the
current natural structure.

## Implementation Notes
Currently this has just been added to the context menu (to expedite not
needing a new icon, etc and also because for puzzles with pip mode, it's
the quickest way that came to mind to figure out which state to apply
to). The UX could definitely use improvement.

## Testing
* Copied target and natural structures
* Pasted alternative structure changing constrained regions with and
without specified start base
* Pasted alternative structure changing unconstrained regions with and
without specified start base

## Related Issues

https://forum.eternagame.org/t/copy-and-paste-secondary-structure-string/2241/3