Add Eternafold + Threshknot folding engine (#691)
## Summary
This commit adds a new folding engine that uses Eternafold for the
initial base pair probability generation, then uses the Threshknot
heuristic (https://arxiv.org/abs/1912.12796) to evaluate the bpp matrix
and generate pairs (including pseudoknots). This engine is being
implemented for use in the Pseudoknot project, with the goal of greatly
speeding up computation time compared to NuPACK+PK.

## Implementation Notes
These heuristic algorithms (Threshknot, Hungarian, etc) take the base
pair probability matrix output of a folding engine and use it to select
base pairs and generate a structure. As such, they can be applied to any
folding engine that can output a base pair probability matrix. We
currently have no way to add a post-fold processing step into the
folding pipeline, so the simplest approach was to create a new folding
engine, `EternaFoldThreshknot`, that uses the Threshknot algorithm in
it's `foldSeqImpl` method. Some notes on this Threshknot implementation:

- There is a filter parameter `theta` that accepts or rejects base pair
probabilities as significant. The original paper uses a `theta` value of
0.3. We uses a value of 0.15 here, which achieves similar correlation
coefficients with the PK50 dataset, while increasing the likelihood of
pseudoknot prediction roughly 4x.
- The original paper gives no details on how the algorithm should handle
equal base pair probabilities (e.g. [1,4] and [1,7] both have a pair
probability of 0.5). This implementation chooses whichever comes first
in the bpp matrix.
- There is a filter step to remove duplicate base pairs should the
algorithm generate them. I've not run into this situation, but it was in
the `arnie` implementation.

## Testing
There is an EternafoldThreshknot test file that checks the engine folds
appropriately, and that it can generate pseudoknots. I also ran the app
locally on the PK100 test puzzle and demonstrated the improved folding
speed.