Fix specs with pseudoknots and melt plot (#693)
## Summary
Previously, pseudoknotted structures were showing "empty" specs in the
specbox and when solutions were uploaded. Additionally, The melt plot
was only (correctly) calculating the % unpaired at 37 degrees, instead
of at each 10-degree interval

## Implementation Notes
The first issue is due to stats inconsistently being stored/retrieved as
having pknots enabled or not. The latter issue is due to getPairs
returning an empty structure when pairs are not available where it was
being checked against being null (due to a refactoring)