Fix offset target structure paste recording unpaired bases incorrectly (#701)
## Summary
Currently, when pasting a target structure with an offset, the result
can wind up being incorrect. This is because unpaired bases in the
pasted structure are also having their pairing partner being adjusted by
the offset, even though it should be left as -1. This could surface in
one of three ways:
1) A user may see the dialog warning about invalid target pairs even
though it appears to be fully valid
2) The visible target structure may be incorrect
3) In either case, the resulting recorded target structure will be
incorrect. Whether or not there are errors beyond bases that should be
`-1` being offset depends on the initial structure.

## Testing
Validated correct behavior in 11627601 with sample sequence and
structure provided by DigitalEmbrace