Support drawing 0-length hairpins (#704)
## Summary
When loading in the PK pilot data with precomputed estimate structures,
some contrafold-derived structures caused a crash in RNALayout. While
impossible in nature, Contrafold can predict 0-length hairpins, which we
did not handle.

## Implementation Notes
I have opted to allocate a "virtual" node for the 0n hairpin like we do
for other hairpins. This should mean that if an NNFE calculation
generates a score for this node it will display, otherwise it will show
as 0.

## Testing
Loaded design 11438186 in puzzle 11318423 (estimate mode)