Surface reason for design browser load failure (#705)
## Summary
A while back, we implemented permission/access rules for viewing
designs. This has commonly caused confusion where a user tries to view
designs and is left with a blank screen, not realizing they aren't
logged in. This change surfaces such errors to the user so they are
aware of what happened.

## Testing
Loaded puzzle as an unauthenticated user