Use new dialog UI for error dialog, don't over-fetch solutions (#706)
## Summary
The main goal of this change is to use the updated dialog styling for
the error dialog. Additionally I noticed that the error dialog in the
design browser when failing to load solutions was being double-triggered
due to two concurent requests. We really have no need to fire off new
refreshes if we're already refreshing (it's just wasteful and
potentially race-condition prone), so this change also ignores refresh
requests when already refreshing

## Implementation Notes
This also moves from the previous "raw" GamePanel approach of
ErrorDialogMode to the mode "modern" GameWindow approach

## Testing
Loaded design browser as an anonymous user