Fix rnapuzzler clash-free layout in feedback mode (#712)
## Summary
Previously when viewing structures in feedback mode with the rnapuzzler
clash-free layout enabled, the tha RNA would be positioned far-offscreen
with a high distance between bases. PoseEdit and PuzzleEdit would
display the same structure correctly.

## Implementation Notes
The issue was in feedback mode, we weren't passing the target structure
to the pose, which we need to reference in rnalyout in order to apply
correct scaling when laying out with rnapuzzler

## Testing
Verified layout correctness with
https://eternagame.org/game/browse/11318423/?filter1=Id&filter1_arg1=11459092&filter1_arg2=11459092