Show stored custom target structure in feedback view mode (#713)
## Summary
Recently, we moved to always uploading cached undoblocks in lab puzzles
in particular so that we could have access to the author's custom target
structure. This was being displayed in pose edit mode (ie, when loaded
with "copy and view") but not in feedback mode (when loaded with "see
results"). This implements showing the custom target structure in
feedback mode

## Implementation Notes
As mentioned in the code, this is rather hacky, but I don't immediately
see a better way to handle this that's relatively simple. We should
revisit this.

## Testing
Added experimental results from an older solution to a test solution on
dev that had fold-data uplaoded and viewed in feedback mode (along with
solutions that did not have data, which also display fine)