Set solution target structure to puzzle default if no custom target available (#714)
## Summary
Previously, if you viewed a solution with cached fold data that included
a custom target structure, then switched to a solution with no cached
fold data, it would retain the custom target of the previous solution.
This is liable to create confusion. Instead, we now reset to the
puzzle's default target if not available.

## Testing
Swapped between solutions on a test puzzle with with and without fold
data/custom target available