Make hotkeys more consistenly available across modes (#715)
## Summary
Previously, there were a number of hotkeys available in the puzzle
solver that shouldve been available in the puzzlemaker and feedback mode
but were not.

## Implementation Notes
This moves handling most hotkeys to the base GameMode class. Note that
the design browser also extends from game mode, and in that case the
hotkeys don't make sense. For that reason, we only "register" the
hotkeys related to the RNA when there are poses available

## Testing
Tried a combination of hotkeys in each mode

## Related Issues
https://forum.eternagame.org/t/puzzlemaker/3086