Support multiple SHAPE conditions in feedback mode (#716)
## Summary
Feedback mode now properly parses multiple SHAPE conditions in solution
experimental data and uses the last-selected SHAPE condition as the
basis of the estimate mode structure. Additionally, condition names in
the dropdown are prefixed by the experimental method (SHAPE or
Degradation)

## Implementation Notes
Still left unhandled is estimate mode when there are no available SHAPE
conditions. In that case, we should remove the estimate mode button (and
switch back to target mode if we're not already), but that
implementation is not completely trivial, that hasn't happened yet, and
everything still works if you don't switch to estimate mode

## Testing
Ensured experimental data and estimate structure displayed and updated
properly given:
* Single shape condition
* Multiple shape conditions
* Multiple shape conditions + degredation conditions