Notify the user if a tutorial requires boosters but they are not unlocked (#762)
## Summary
Currently, if a tutorial highlights the booster button or waits for it
to be clicked, a user
who does not have boosters unlocked will be left with nothing
highlighted or stuck. This lets
them know that this is a feature they need to unlock.

## Implementation Notes
While there is some argument to be had about whether it is better to
notify the user immediately
on starting the puzzle, inserting the notification when the UI element
is actually requested
is by far the most straightforward to implement, so I have done that at
least for now.

## Testing
Tried a tutorial with booster commands while logged out