Ensure all actions that require folding to complete actually wait (#768)
## Summary
Improves code clarity and robustness to race conditions related to
asynchronous folding

## Implementation Notes
Unless we are in synchronous script mode, whenever a folding operation
(or really, anything that requires state to be resynced/an undo state to
be created) is triggered, the update is actually queued and run on the
next game tick/update cycle - and could in fact be an asynchronous
operation which does not complete for some time, even though the game
continues to run (though under a UI lock). I've gone through and audited
instances where we appear to perform actions immediately after we
trigger a fold without waiting for the fold to complete.

I've now verified that we now only trigger async behaviors in callbacks,
and if not explicitly awaited, is the very last thing to happen in the
callback. The only instance that I did not address is folding engine
changes, which is triggered by a FolderSwitcher method which updates its
current state and then notifies the mode of a change via a signal.
Ideally we would be able to await a manual engine change and the folder
switcher would be UI only, but this requires some more time than I have
right now and the behavior is fine.

Breaking down the specific changes:
* `poseEditByTarget` now registers a poseOp it can use to return a
Promise that callers can `await` on instead of having to do that
themselves
* Instead of `pasteSequence` being handled in pose2D (with the update
being propagated by the `poseEditCallback` like happens with UI
mutations), we now handle this directly in the modes themselves,
directly calling `poseEditByTarget` and marking `pasteSequence` async.
This also nicely avoids some indirection. Also, note that previously
we've called `pasteSequence` in a loop over every pose - this is wholly
unnecessary as `poseEditByTarget` already takes care of syncing the
updates from the pose that was changed (indicated by `targetIndex`) to
other poses.
* `publicStartCountdown` is never used, so removed
* `setSolutionTargetStructure` now assumes callers properly wait for
folding to complete before calling it
* `showSolution` is now written more cleanly as an async function, waits
for folding to complete before making further undoBlock modifications,
and doesnt go through the extra pasteSequence logic (this clarifies
where the poseEditByTarget happens since we perform multiple state
updates, and there's no need to do the extra validation logic since this
is an already submitted sequence)
* We start loading the 3D file if present before folding completes, but
wait until it completes to load the 3D window (it does rely on fold
completion, but we'd never seen a situation before where we would get
this file back before the fold completed)
* RScript does can be initialized before folding completes (it's already
designed this way), so we move that before our initial fold
* 3D and initial target loading are moved into a PoseOp so that it
happens after the initial fold but before user interaction
* `setPuzzleEpilog`/`_poseEditByTargetCb` was inlined instead of being a
magic state variable
* `showNextSolution` now has the `showSolution` call (which actually
triggers the state update) at the very end (whether or not eg the
solution sidebar shows up before or after doesnt really matter, but it
is now consistent)
* `updateHighlightsAndScores` didn't appear to actually be doing
anything. It appears that the bug that caused it to be introduced was
separately resolved by the highlight being cleared elsewhere eg the
update loop (in the same commit)
* Pose2D.puzzleLocks never returns undefined, so it no longer says it
can