Ensure structure operations are always performed with pseudoknots where required (#775)
## Summary
#754 changed the behavior of secondary structure parsing to error when
encountering an unknown or unsupported character. This allows for
increased robustness, and ensuring that we never silently discard
information. However, some areas of the code, such as drawing secondary
structure visualizations in the design browser, implicitly used this
behavior to strip pseudoknots. This change makes the pseudoknot
parameter for these functions always required and audits all times
they're called to ensure that we always process with pseudoknots when
needed,

## Implementation Notes
* `getParenthesis` now takes an object rather than ordered arguments.
Practically speaking this is because it made the refactoring
semi-automated to be able to remove the seq parameter when not needed,
though I also think this winds up being a little more readable.
* In folders, even if they (or some subset of methods) can't handle
pseudoknots I explicitly use the `pseudoknots` parameter or allocate a
`pseudoknots` constant and use that so that: 1) it is easier to trace
that we expect not to work with pknots 2) if the code is copy-pasted to
an area where it can work with pknots, it's clearer that such code needs
updating (or is already handled by a parameter) and 3) if we add pknot
support later (eg, in additional areas nupack supports but we do not),
it's clearer to remember that needs to change (or is already handled by
a parameter)
* There are a number of areas where this is basically a noop because the
codepaths are never currently called with pseudoknots enabled (eg due to
no engines supporting pknots and ligands/aptamers/oligos), but I handle
them now so that we don't forget to later
* A number of eternascript functions now handle pknots more correctly.
That said, get_defect is rarely used, pairing_probabilities currently
never returns anything different in pseudoknotted engines anyways, and
no pseudoknot engine currently handles energy calculation properly
either
* Resetting the target structure (via the paste structure dialog, the
invalid-pairings-with-custom-target dialog, or loading a solution
without target structures) now correctly includes pseudoknots - though I
don't believe we've had a puzzle yet that both has a pseudoknotted
target and allows for custom target structures