Don't lock script methods during folding if called from script constraints (#778)
Script constraints may call various methods that we've started to lock
during folding - most obviously, constraints probably need to access the
current solution to know whether it passes the constraint or not.
Unfortunately, we can't really limit this to a set of reasonable methods
(eg, just getters) because there are a bunch of constraints that
currently hack in additional functionality this way.

It is still possible that there are some edge cases (eg, a constraint
that does some very fast async work which yields the special external
interface context and then goes and uses one of these methods before the
regular lock releases), but I'm going to leave it like this to attempt
being at least a bit safer and hope it doesn't cause issues (and if it
does, it's likely it would wind up having other issues _anyways_).