Fix set_sequence_string being overridden with a version without async handling (#779)
When executing a booster, we provide it with a context which overrides
the set_sequence_string method so that "paint tool" scripts can act
directly on the pose during the paint process (eg, allowing a temporary
mutation during a mouse move, and avoiding conflicts between the
"UI-based" mutation process which was initiated and "script-based"
mutation, both of could otherwise trigger fold/undo operations).

However, this method is always overridden and did not change when we
changed the base implementation to accommodate async folding (which
means the additional checks would not be present when this method is
called from a synchronous script). This instead only overrides the
method for paint tool scripts, allowing the base implementation to be
directly used in all other cases