Fix UI locks not applying to info sidebar (#780)
## Summary
Currently, when the UI is locked, the sidebar can still be interacted
with. This is now prevented.

## Implementation Notes
* This would be trivially solved if the sidebar could be under dialogs
(where it realistically belongs). However, due to it containing HTML
content that would render over other dialogs, we do not allow this.
Therefore I instead draw a second lock dialog on top of the sidebar
layer and mask it to where the sidebar is.
* I also handle "modal dialogs", which should have the same effect of
locking the UI. However in this case, I also hide the sidebar entirely
since otherwise you may not be able to see/interact with part of the
likely statically-positioned dialog if the screen is small enough.
* The use of `any` here is due to typescript generic inference
limitations. This is type safe from the consumer level, it just leaves
us open to misusing the return of the dialog closure, which we should
never touch anyways.
* While I was debugging I also noticed dialogLayer and sidebayLayer had
the wrong container names (surfaced in the pixi devtools). This is now
resolved.